# [SS] Use buildbuddy repo for manual performance test for more realistic results