--- a/enterprise/server/remote_execution/containers/firecracker/BUILD
+++ b/enterprise/server/remote_execution/containers/firecracker/BUILD
@@ -83,7 +83,10 @@ genrule(
 go_test(
     name = "firecracker_test",
     timeout = "long",
-    srcs = ["firecracker_test.go"],
+    srcs = [
+        "firecracker_performance_test.go",
+        "firecracker_test.go",
+    ],
     exec_properties = {
         "test.Pool": "bare",
         "test.use-self-hosted-executors": "true",
@@ -144,7 +147,10 @@ go_test(
 go_test(
     name = "firecracker_test_blockio",
     timeout = "long",
-    srcs = ["firecracker_test.go"],
+    srcs = [
+        "firecracker_performance_test.go",
+        "firecracker_test.go",
+    ],
     args = [
         "--executor.firecracker_enable_vbd=true",
         "--executor.firecracker_enable_merged_rootfs=true",