I just changed the test to build the buildbuddy repo instead of gazelle, and moved it to its own file.

**Updated results:**

(remote_snapshot_sharing=true) Bazel build on a clean runner took 23m4.322567448s.
(remote_snapshot_sharing=true) Bazel build on a recycled runner (100% locally cached) took 13.219176893s.
(remote_snapshot_sharing=true) Bazel build (30% locally cached) took 25.386624531s. 70% artifacts fetched remotely.
(remote_snapshot_sharing=true) Bazel build (0% locally cached) took 58.381049801s. 100% artifacts fetched remotely.

(remote_snapshot_sharing=false) Bazel build on a clean runner took 22m34.195032295s.
(remote_snapshot_sharing=false) Bazel build on a recycled runner (100% locally cached) took 14.297437927s.
(remote_snapshot_sharing=false) Bazel build (30% locally cached) took 22m36.815208793s. Could not start from snapshot, had to prepare clean runner.
(remote_snapshot_sharing=false) Bazel build (0% locally cached) took 22m56.842704256s. Could not start from snapshot, had to prepare clean runner.


======= More Detailed Breakdowns =======

(remote_snapshot_sharing=true) Pausing a new VM took 1.369853545s.
(remote_snapshot_sharing=false) Pausing a new VM took 1.036052469s.
(remote_snapshot_sharing=true) Pausing a VM that had started from a snapshot and had no non-metadata changes took 1.013146486s.
(remote_snapshot_sharing=false) Pausing a VM that had started from a snapshot and had no non-metadata changes took 894.046684ms.
(remote_snapshot_sharing=true) Pausing a VM that had started from a snapshot and had execution-related changes took 1.093709081s.
(remote_snapshot_sharing=false) Pausing a VM that had started from a snapshot and had execution-related changes took 1.550299784s.
(remote_snapshot_sharing=true) Unpausing a VM that is fully cached in filecache took 2.109292876s.
(remote_snapshot_sharing=false) Unpausing a VM that is fully cached in filecache took 1.97428592s.
(remote_snapshot_sharing=true) Unpausing a VM where ~30% of artifacts were evicted from filecache and must be fetched remotely. Took 2.492413683s.
(remote_snapshot_sharing=true) Unpausing a VM where all artifacts were evicted from filecache and must be fetched remotely. Took 2.790781031s.